+++ Changelog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 -   Add MT_DEBUG_INTERVAL setting
+-   Send SIGKILL if SIGTERM does not cause test to exit
 
 ## [1.6.0] 2021-01-06